requirements.txt
discord
discord.py
python-dotenv
Pillow
requests
psycopg2
paramiko
yt_dlp
urllib3
spotipy